requirements.txt
amqp==2.6.1
asgiref==3.5.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
Brotli==1.0.9
celery==4.4.7
cffi==1.15.0
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
crispy-bootstrap5==0.6
cssselect2==0.6.0
Deprecated==1.2.13
Django==4.0.3
django-crispy-forms==1.14.0
djangorestframework==3.13.1
flower==0.9.7
fonttools==4.32.0
html5lib==1.1
humanize==4.0.0
kombu==4.6.11
packaging==21.3
Pillow==9.1.0
prometheus-client==0.8.0
prompt-toolkit==3.0.28
pycparser==2.21
pydyf==0.1.2
pyparsing==3.0.7
pyphen==0.12.0
pytz==2022.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
tornado==6.1
vine==1.3.0
wcwidth==0.2.5
weasyprint==54.3
webencodings==0.5.1
wrapt==1.14.0
zopfli==0.2.1
